Remove bpf from encoders and unneeded functions
