+++ encoder/csv_test.go
@@ -37,3 +37,16 @@ bar,udp,[::1]:60051,[::1]:60052
 		t.Fatalf("Unexpected output: wanted \"%s\", found \"%s\"", expOut, w.String())
 	}
 }
+
+var netFiles0 = []lookup.NetFile{
+	{"foo", newUDPAddr("192.168.0.61:54104"), newUDPAddr("52.94.218.7:443")},
+	{"bar", newUDPAddr("[::1]:60051"), newUDPAddr("[::1]:60052")},
+}
+
+func newUDPAddr(address string) net.Addr {
+	addr, err := net.ResolveUDPAddr("udp", address)
+	if err != nil {
+		panic(err)
+	}
+	return addr
+}
